Update test documentation and name
