+++ convex/__tests__/userQuests.test.ts
@@ -2,8 +2,8 @@ import { convexTest } from "convex-test";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { vi } from "vitest";
 import { api } from "../_generated/api";
-import { createUser } from "../helpers";
 import schema from "../schema";
+import { createTestUser } from "../test-helpers";
 import { modules } from "../test.setup";
 
 const UPDATE_TIMESTAMP = Date.now();
@@ -21,7 +21,7 @@ describe("userQuests", () => {
   describe("getAll", () => {
     it("should return all user quests", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       await t.run(async (ctx) => {
         // Create quests
@@ -65,7 +65,7 @@ describe("userQuests", () => {
 
     it("should not return deleted quests", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       await t.run(async (ctx) => {
         const questId = await ctx.db.insert("quests", {
@@ -93,7 +93,7 @@ describe("userQuests", () => {
   describe("count", () => {
     it("should return count of user quests", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       await t.run(async (ctx) => {
         const questId = await ctx.db.insert("quests", {
@@ -120,7 +120,7 @@ describe("userQuests", () => {
   describe("create", () => {
     it("should create a user quest", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await asUser.run(async (ctx) => {
         return await ctx.db.insert("quests", {
@@ -146,7 +146,7 @@ describe("userQuests", () => {
 
     it("should set a default status", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await asUser.run(async (ctx) => {
         return await ctx.db.insert("quests", {
@@ -167,7 +167,7 @@ describe("userQuests", () => {
 
     it("should throw if quest already exists", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return await ctx.db.insert("quests", {
@@ -191,7 +191,7 @@ describe("userQuests", () => {
   describe("setStatus", async () => {
     it("should update quest status", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return ctx.db.insert("quests", {
@@ -217,7 +217,7 @@ describe("userQuests", () => {
 
     it("should throw error if status is invalid", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return ctx.db.insert("quests", {
@@ -242,7 +242,7 @@ describe("userQuests", () => {
 
     it("should properly manage timestamps through status transitions", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return ctx.db.insert("quests", {
@@ -290,7 +290,7 @@ describe("userQuests", () => {
 
     it("should remove completedAt when status changed away from complete", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return ctx.db.insert("quests", {
@@ -334,7 +334,7 @@ describe("userQuests", () => {
   describe("deleteForever", () => {
     it("should delete a user quest", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       const questId = await t.run(async (ctx) => {
         return await ctx.db.insert("quests", {
@@ -364,7 +364,7 @@ describe("userQuests", () => {
   describe("getByCategory", () => {
     it("should return quests grouped by category", async () => {
       const t = convexTest(schema, modules);
-      const { asUser, userId } = await createUser(t);
+      const { asUser, userId } = await createTestUser(t);
 
       await t.run(async (ctx) => {
         const quest1Id = await ctx.db.insert("quests", {
